backend/functions/audio_processor/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-firestore==2.*
google-cloud-secret-manager==2.*
requests==2.*
openai==1.*
google-cloud-pubsub==2.*
